requirements.txt
pyTelegramBotAPI
rtoml>=0.9.0
pathlib>=1.0.1
rich>=13.3.5
loguru>=0.5.3
httpx>=0.23.3
pydantic>=1.10.4
beautifulsoup4>=4.11.1
python-dotenv
elara
redis>=4.5.4
motor
Pillow
shortuuid